requirements.txt
PySide6>=6.6.0,<7.0.0
Pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0
edge-tts>=6.1.0
pywin32>=306
PyYAML>=6.0,<7.0
opencv-python>=4.8.0,<5.0.0
mediapipe>=0.10.0,<1.0.0
mss>=9.0.0
paddleocr>=2.7.0
openai>=1.0.0
httpx>=0.24.0
